+++ libs/application/templates/health-insurance/src/fields/index.ts
@@ -1,2 +1,2 @@
-export { default as Review } from './Review/Review'
+export { default as Review } from './Review'
 export { default as ConfirmCheckbox } from './ConfirmCheckbox/ConfirmCheckbox'
